+++ resources/grunt/uglify.js
@@ -21,6 +21,7 @@ module.exports = {
                 "build/js/paypal-frontend-standard-payment-controller.js",
                 "build/js/paypal-frontend-acdc-payment-controller.js",
                 "build/js/paypal-frontend-googlepay-payment-controller.js",
+                "build/js/paypal-frontend-applepay-payment-controller.js",
                 "build/js/paypal-frontend-express-payment-controller.js"
             ]
         }