+++ src/Core/PatchRequestFactory.php
@@ -54,7 +54,12 @@ class PatchRequestFactory
         $order = oxNew(Order::class);
         $order->load($orderId);
 
-        $patches = [];
+        $patches = array_values(
+            array_filter([
+                $this->getAmountPatch(),
+                $this->getPurchaseUnitsPatch()
+            ])
+        );
 
         if ($deliveryId && $deliveryAddress->load($deliveryId)) {
             $patches[] = $this->getShippingNamePatch($deliveryAddress);