+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [0007821](https://bugs.oxid-esales.com/view.php?id=7821): use finalizeOrder for all payments again
 - fix PUI-EMail-Handling
 - [0007821](https://bugs.oxid-esales.com/view.php?id=7821): use finalizeOrder for all payments again
-- use PayPal-Client v3.0.18
+- use PayPal-Client v3.0.19
 - Add handling for worst case: When the order is completed and the customer clicks around, the customer is redirected to the thank you page
+- Add new webhook 'VAULT.PAYMENT-TOKEN.CREATED'
 
 ### FIX